requirements.txt
Flask==3.0.0
pandas==2.1.4
Werkzeug==3.0.1
python-dateutil==2.8.2
numpy==1.26.2
gunicorn==21.2.0
python-dotenv==1.0.0
